Implement revert with keep original option

+++ versioncontrol/actions/git_revert.py
@@ -158,27 +158,28 @@ def undo_files(path: str, files: list[str], entry_id: str, channel_id: str):
             logging.info(str(e))
             ui.show_error("Undo Failed", str(e))
 
-def restore_files(path: str, files: list[str], entry_id: str, channel_id: str):
+def restore_files(path: str, files: list[str], entry_id: str, channel_id: str, keep_original: bool):
     ui = ap.UI()
     if len(files) == 0:
         ap.UI().show_success("No Files Selected")
 
     try:
-        progress = ap.Progress("Restoring Files", show_loading_screen=True)
         repo = GitRepository.load(path)
         repo_root = repo.get_root_path()
- 
-         # Check if any file to revert is locked by an application
+
+        progress = ap.Progress("Restoring Files", show_loading_screen=True)
+
+        # Check if any file to revert is locked by an application
         relative_selected_paths = set()
         for path in files:
             relpath = os.path.relpath(path, repo_root)
             relative_selected_paths.add(relpath.replace("\\", "/"))
-            if not utility.is_file_writable(path):
+            if not keep_original and not utility.is_file_writable(path):
                 error = f"error: unable to unlink '{relpath}':"
                 if not git_errors.handle_error(error):
                     ui.show_info("Could not restore files", f"A file is not writable: {relpath}", duration=6000)
                 return True
-            
+                
         changes = repo.get_changes_for_changelist(entry_id)
         for deleted_file in changes.deleted_files:
             if deleted_file.path in relative_selected_paths:
@@ -188,10 +189,17 @@ def restore_files(path: str, files: list[str], entry_id: str, channel_id: str):
             ui.show_success("Nothing to restore")
             return
 
-        repo.restore_files(list(relative_selected_paths), entry_id)
+        changes = repo.get_all_pending_changes()
+        pending_changes_count = changes.size()
 
-        changes = repo.get_pending_changes(True)
-        restored_files = changes.size()
+        repo.restore_files(list(relative_selected_paths), entry_id, keep_original)
+
+        changes = repo.get_all_pending_changes()
+        if keep_original:
+            restored_files = len(relative_selected_paths)
+        else:
+            restored_files = changes.size() - pending_changes_count
+            
         if restored_files == 1:
             ui.show_success("Restore Successful", "One file has been restored")
         elif restored_files > 1:
@@ -207,6 +215,20 @@ def restore_files(path: str, files: list[str], entry_id: str, channel_id: str):
         if not git_errors.handle_error(e):
             logging.info(str(e))
             ui.show_error("Restore Failed", str(e))
+
+def async_wrapper(func, dialog, *args, **kwargs):
+    dialog.close()
+    ap.get_context().run_async(func, *args, **kwargs)
+
+def show_restore_files_dialog(path: str, files: list[str], entry_id: str, channel_id: str):
+    dialog = ap.Dialog()
+    dialog.title = "Restore Files"
+    dialog.icon = ":/icons/restore.svg"
+    dialog.add_text("This might overwrite files in your project.")
+    dialog.add_text("Do you want to keep the original files?")
+    dialog.add_info("A file <b>level.umap</b> will be restored as <b>level_restored.umap</b>.")
+    dialog.add_button("Overwrite Files", primary=False, callback=lambda d: async_wrapper(restore_files, d, path, files, entry_id, channel_id, False)).add_button("Keep Original Files", primary=False, callback=lambda d: async_wrapper(restore_files, d, path, files, entry_id, channel_id, True))
+    dialog.show()
 
 def on_timeline_detail_action(channel_id: str, action_id: str, entry_id: str, ctx: ap.Context):
     ui = ap.UI()
@@ -230,24 +252,9 @@ def on_timeline_detail_action(channel_id: str, action_id: str, entry_id: str, ct
         finally:
             return True
     if action_id == "gitrestorecommitfiles":
-        try:
-            path = get_repo_path(channel_id, ctx.project_path)
-            repo = GitRepository.load(path)
-            if not repo: return
-
-            if repo.has_pending_changes(True):
-                ui.show_info("Cannot restore files", "You have changed files. Commit them and try again")
-                return True
-            else:
-                restore_files(path, ctx.selected_files, entry_id, channel_id)
-                ap.refresh_timeline_channel(channel_id)
-
-        except Exception as e:
-            if not git_errors.handle_error(e):
-                logging.info(str(e))
-                ui.show_error("Restore Failed", str(e))
-        finally:
-            return True
+        path = get_repo_path(channel_id, ctx.project_path)
+        show_restore_files_dialog(path, ctx.selected_files, entry_id, channel_id)
+        return True
         
     if action_id == "gitrevertcommitfiles":
         try:
